Style improvements and missing .gitignores
