--- app.test.js
@@ -1,5 +1,0 @@
-const { hello } = require("./app");
-
-test("hello greets properly", () => {
-  expect(hello("Kaholo")).toBe("hello Kaholo");
-});
