+++ jwtFile.js
@@ -16,12 +16,15 @@ async function createJWTFile({ sourceDirectory, jwtKey }) {
       CREATE_APPEND_FAIL_IF_EXISTS_FLAGS,
     );
     await writeFile(fileHandle, jwtKey);
-    await fileHandle.close();
   } catch (error) {
     if (error.code === NO_SUCH_FILE_OR_DIRECTORY_SYSTEM_ERROR) {
       throw new Error(`Directory ${sourceDirectory} does not exist`);
     } else {
       throw error;
+    }
+  } finally {
+    if (fileHandle) {
+      await fileHandle.close();
     }
   }
 }
